PUT agora funcional, diversas melhorias no front e refatorações visando um codigo limpo

+++ loja-de-veiculos-spring/src/main/java/com/fag/lojadeveiculos/presentations/controllers/VeiculoController.java
@@ -16,7 +16,6 @@ import org.springframework.web.bind.annotation.RestController;
 import com.fag.lojadeveiculos.domain.dto.VeiculoDTO;
 import com.fag.lojadeveiculos.domain.entities.VeiculoBO;
 import com.fag.lojadeveiculos.domain.mapper.VeiculoMapper;
-import com.fag.lojadeveiculos.domain.repositories.IVeiculoImagemRepository;
 import com.fag.lojadeveiculos.domain.repositories.IVeiculoRepository;
 
 @RestController
@@ -26,44 +25,29 @@ public class VeiculoController {
 
     @Autowired
     private IVeiculoRepository veiculoRepository;
-    @Autowired
-    private IVeiculoImagemRepository veiculoImagemRepository;
 
     @GetMapping
     public List<VeiculoDTO> getAll() {
-        //List<VeiculoImagemBO> imagems = veiculoImagemRepository.findAll();
-
         List<VeiculoDTO> veiculos = veiculoRepository.findAll().stream()
                 .map(veiculo -> VeiculoMapper.toDTO(veiculo)).toList();
-
-        // veiculos.stream().forEach(veiculo -> veiculo.setListVeiculoImagem(
-        //         imagems.stream().filter(imagem -> imagem.getVeiculo().getId() == veiculo.getId()).toList()));
-
         return veiculos;
     }
 
     @PostMapping
-    public void saveVeiculo(@RequestBody VeiculoDTO veiculoDTO){
-        veiculoRepository.save(VeiculoMapper.toBO(veiculoDTO));
-
-        //List<VeiculoImagemBO> listVeiculoImagem = veiculoDTO.getListVeiculoImagem();
-        //if(listVeiculoImagem != null && !listVeiculoImagem.isEmpty())
-          //  listVeiculoImagem.forEach(veiculoImagem -> veiculoImagemRepository.save(veiculoImagem));
-            
-    }
-
-    @DeleteMapping(value = {"/{id}"})
-    public void deleteVeiculo(@PathVariable Long id){
-        veiculoRepository.deleteById(id);
-    }
-
-    @PutMapping
-    public void updateVeiculo(@RequestBody VeiculoDTO veiculoDTO){
+    public void saveVeiculo(@RequestBody VeiculoDTO veiculoDTO) {
         VeiculoBO veiculo = VeiculoMapper.toBO(veiculoDTO);
         veiculoRepository.save(veiculo);
     }
 
-    public VeiculoBO getVeiculoById(Long idVeiculo){
-        return veiculoRepository.findById(idVeiculo).orElse(null);
+    @DeleteMapping(value = { "/{id}" })
+    public void deleteVeiculo(@PathVariable Long id) {
+        if (veiculoRepository.existsById(id))
+            veiculoRepository.deleteById(id);
+    }
+
+    @PutMapping
+    public void updateVeiculo(@RequestBody VeiculoDTO veiculoDTO) {
+        VeiculoBO veiculo = VeiculoMapper.toBO(veiculoDTO);
+        veiculoRepository.saveAndFlush(veiculo);
     }
 }
